AI-Agents-with-MongoDB/01-Building-AI-Agents-with-MongoDB/requirements.txt
boto3
langchain-aws
langchain-openai
pymongo
voyageai
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
langchain-aws
boto3
notebook
langchain-tavily
wikipedia
trustcall
langgraph-cli[inmem]
python-dotenv
datasets
pandas